+++ src/engine/teleport_check.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   teleport2.c                                        :+:      :+:    :+:   */
+/*   teleport_check.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mott <mott@student.42heilbronn.de>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/29 14:17:30 by mott              #+#    #+#             */
-/*   Updated: 2024/07/01 13:27:23 by mott             ###   ########.fr       */
+/*   Updated: 2024/07/01 16:02:25 by mott             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/cub3D.h"
 
-static void	teleport_order_x(t_game *game, t_coords dest, double x)
+static void	teleport_order_x(t_game *game, t_coords dest, float x)
 {
 	if (x < 0)
 	{
@@ -38,7 +38,7 @@ static void	teleport_order_x(t_game *game, t_coords dest, double x)
 	}
 }
 
-static void	teleport_order_y(t_game *game, t_coords dest, double y)
+static void	teleport_order_y(t_game *game, t_coords dest, float y)
 {
 	if (y < 0)
 	{
@@ -64,7 +64,7 @@ static void	teleport_order_y(t_game *game, t_coords dest, double y)
 	}
 }
 
-void	teleport_check(t_game *game, double x, double y)
+void	teleport_check(t_game *game, float x, float y)
 {
 	if ((int)(game->player->pos.x + x) >> 6 == game->map->portal[0].x
 		&& (int)(game->player->pos.y) >> 6 == game->map->portal[0].y)
